fix: simplify interception by processing requests directly
Fixes a rare issue where the NZB file is not intercepted because the redirected request receives a new request ID.
Since we use the same regular expression for the redirection rules and the isURLTracked check, we can safely assume that every tracked request will be blocked. Requests from the extension itself are excluded by the tabID check and the x-nzbdonkey parameter check.